requirements.txt
boto3>=1.34.0
Pillow>=10.0.0
python-dotenv>=1.0.0
pandas>=2.0.0